Rename CC3220 preset to CC3220SF
