+++ make/mt-os/mt-os.mk
@@ -10,7 +10,7 @@ XI_CONST_PLATFORM_ARM := arm-linux
 XI_CONST_PLATFORM_WMSDK := wmsdk
 XI_CONST_PLATFORM_MICROCHIP := microchip
 XI_CONST_PLATFORM_CC3200 := cc3200
-XI_CONST_PLATFORM_CC3220 := cc3220
+XI_CONST_PLATFORM_CC3220SF := cc3220sf
 XI_CONST_PLATFORM_STM32FX := stm32fx
 XI_CONST_PLATFORM_STM32FX_NUCLEO_WIFI := stm32fx_nucleo_wifi
 
@@ -39,8 +39,8 @@ ifneq (,$(findstring $(XI_CONST_PLATFORM_WMSDK),$(TARGET)))
   XI_CONST_PLATFORM_CURRENT := $(XI_CONST_PLATFORM_WMSDK)
 endif
 
-ifneq (,$(findstring $(XI_CONST_PLATFORM_CC3220),$(TARGET)))
-  XI_CONST_PLATFORM_CURRENT := $(XI_CONST_PLATFORM_CC3220)
+ifneq (,$(findstring $(XI_CONST_PLATFORM_CC3220SF),$(TARGET)))
+  XI_CONST_PLATFORM_CURRENT := $(XI_CONST_PLATFORM_CC3220SF)
 endif
 
 ifneq (,$(findstring $(XI_CONST_PLATFORM_CC3200),$(TARGET)))
