+++ src/bsp/platform/cc3220sf/xi_bsp_mem_cc3220sf.c
