+++ src/bsp/platform/cc3220sf/xi_bsp_rng_cc3220sf.c
