+++ src/bsp/platform/cc3220sf/xi_bsp_time_cc3220sf_sntp.c
@@ -11,7 +11,7 @@
  *  sntp_task.c                                                               *
  *                                                                            *
  ******************************************************************************/
-#include "xi_bsp_time_cc3220_sntp.h"
+#include "xi_bsp_time_cc3220sf_sntp.h"
 #include "simplelink.h"
 
 /******************************************************************************
